requirements.txt
matplotlib==3.5.3
numpy==1.21.5
Pillow==11.2.1
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.13.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
